Update stack.cpp with Fernando Apesteguía's patch for compiling on 64bit.
git-svn-id: https://svn.code.sf.net/p/kidbasic/code/trunk@524 deda80ff-9d1e-0410-aade-8f0ad8a54eb3

+++ Stack.cpp
@@ -53,7 +53,7 @@ Stack::debug()
 int Stack::height()
 {
 	// return the height of the stack in elements
-	return ((unsigned int) top - (unsigned int) bottom)/sizeof(stackval);
+	return ((unsigned int) (top - bottom)/sizeof(stackval));
 }
 
 void
